+++ melt_the_code_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Implemented the useLinkOpener() use case function.
+
 ## 0.1.0
 
 * Initial release.